Move BaseTest to fix import issues
